requirements.txt
streamlit
pandas
numpy
yfinance
tweepy
textblob
vaderSentiment
python-dotenv
plotly